Refactor search logic into useFridgeSearch hook
